requirements.txt
Django==2.2.19
Pillow==9.2.0
pytz==2022.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
sqlparse==0.4.3